fix(SRS-MD-005): make the admission guard a CLOSED SET, not a checklist
Round 4: two BLOCKs, both the same guard, plus three warns.

BLOCK — the admission-point discovery has now been wrong twice, both times
because it enumerated something I could think of. First it discovered sites by
'takes a RestartWindowGate' (circular: a path that SKIPS the port is exactly
what must be caught). Round 3 replaced that with two literal effect FORMS, and
the reviewer walked straight past both with
`subscribers.entry(k).or_default().push(..)`, which opens an upstream line just
as well. A guard shaped like a checklist cannot catch the arm nobody added.

Discovery is now a closure argument over the TYPE's surface rather than over
remembered syntax: every public `&mut self` method on
ConsolidatedSubscriptionRegistry — the only way to change the consolidated set
at all, whatever expression it uses — plus every public minter of the acceptance
envelope, which covers the manager (a separate type holding no registry field).
A new admission path must be one or the other. `unsubscribe` is a declared
per-item exemption with a stated reason (releasing a line is always safe, and
blocking it would strand strategies), and a STALE exemption fails too. Four
bypass shapes are pinned as regression tests, including the reviewer's.

WARN — a ProbeFailed from THIS host (exhausted descriptors, denied permission)
collapsed to Unreachable and paged as 'IB Gateway is unreachable'. It still
fails closed, which is right, but it misattributes — and the module claimed the
operator could see the probe's own failure, which they could not.
`last_outcome()` now retains it, the CLI prints it, the claim is scoped to what
holds, and routing it into the alert is declared deferred (ConnectivityEvent's
field set is pinned by ERR-2, which forbids transport detail).

WARN — the catalogued keys validated and changed NOTHING: .env.example, the
config README and the catalogue all describe them as controlling the suspension
window while the binary read compiled-in constants. A documented knob that moves
no behaviour is a lie with a fuse, and this one would have been discovered
during a restart. Both duration keys now seed the CLI's defaults, with flags
winning over them; malformed, empty and non-positive values are typed errors,
not silent fallbacks; NotUnicode is not treated as absent. ATP_IB_RESTART_ET
stays Python-resolved (it needs the DST calendar) and its wiring is declared
deferred with the daemon that will own it.

WARN — the reconnect ledger grew without bound on a sustained outage. Bounded
ring of 256 with a dropped-count, and the TOTAL stays exact: truncating the
count would under-report an outage, the wrong direction.

Verified: cargo test --workspace (176 suites); clippy -D warnings; fmt clean;
connectivity_check; architecture_check; pytest 'not integration and not e2e'
(5361 passed); domain suite 34 passed; L5 fault injection (8 passed).

+++ crates/atp-orchestrator/src/restart_window_scenario.rs
@@ -158,7 +158,7 @@ pub struct RestartWindowEvidence {
     /// Wire attempts. A blocked submission must create ZERO IB orders — that is
     /// what proves the refusal happened at the gate and not at the broker.
     pub ib_orders_created: u32,
-    pub reconnects: usize,
+    pub reconnects: u64,
     pub events_recorded: usize,
     pub event_scheduled_restart: Option<bool>,
     /// SYS-75(a), market-data half.
